git-svn-id: svn://svn.icms.temple.edu/lammps-ro/trunk@8625 f3b2605a-c512-4ea7-a41b-209d697bcdaa

+++ src/REPLICA/prd.cpp
@@ -482,7 +482,7 @@ void PRD::dynamics()
   update->integrate->setup();
   // this may be needed if don't do full init
   //modify->addstep_compute_all(update->ntimestep);
-  int ncalls = neighbor->ncalls;
+  bigint ncalls = neighbor->ncalls;
 
   timer->barrier_start(TIME_LOOP);
   update->integrate->run(t_event);
